Refactor models with modular structure
- Split large model files into smaller, focused modules
- Create separate modules for Gear and Patch functionality
- Add modular exports via index.ts files
- Update CLAUDE.md with module structure guidelines
- Add MIGRATION.md documentation for gradual transition
- Add PR review documentation
