+++ MIGRATION.md
@@ -1,0 +1,64 @@
+# Migration Plan for Modular Structure
+
+This document outlines the step-by-step process to migrate from the current monolithic model structure to the new modular one. The goal is to minimize disruption while improving maintainability.
+
+## Overview of Changes
+
+We've refactored the large `Gear.ts` and `Patch.ts` files into smaller, more focused modules:
+
+### Gear Module Structure
+- `lib/models/gear/types.ts` - Interface definitions
+- `lib/models/gear/Gear.ts` - Core Gear class
+- `lib/models/gear/processing.ts` - LLM processing functionality
+- `lib/models/gear/connections.ts` - Connection management
+- `lib/models/gear/examples.ts` - Example inputs management
+- `lib/models/gear/index.ts` - Public exports
+
+### Patch Module Structure
+- `lib/models/patch/types.ts` - Interface definitions
+- `lib/models/patch/Patch.ts` - Core Patch class
+- `lib/models/patch/nodes.ts` - Node management
+- `lib/models/patch/edges.ts` - Edge management
+- `lib/models/patch/reactflow.ts` - ReactFlow integration
+- `lib/models/patch/utils.ts` - Static utilities
+- `lib/models/patch/index.ts` - Public exports
+
+## Migration Approach
+
+### Phase 1: Prepare for Migration
+1. Create the new directory structure (completed)
+2. Implement the new modular files (completed)
+3. Create the index.ts files for proper exports (completed)
+
+### Phase 2: Gradual Migration
+1. Create new module index.ts to export all components from the new structure
+2. Keep the original Gear.ts and Patch.ts files temporarily as pass-through modules
+3. Update the original files to re-export from the new modules
+4. Verify application functionality with the redirected imports
+
+### Phase 3: Update Import References
+1. Identify all import references to the original files
+2. Update imports to use the new modular structure:
+   - Change `import { Gear, ... } from '../models/Gear'` to `import { Gear, ... } from '../models/gear'`
+   - Change `import { Patch, ... } from '../models/Patch'` to `import { Patch, ... } from '../models/patch'`
+3. Test each component after updating imports
+
+### Phase 4: Cleanup
+1. Once all imports are updated, remove the original Gear.ts and Patch.ts files
+2. Run comprehensive tests to ensure all functionality works properly
+3. Update documentation to reflect the new structure
+
+## Benefits of New Structure
+
+1. **Improved Maintainability**: Each file has a clear, single responsibility
+2. **Better Organization**: Related functionality is grouped together
+3. **Reduced File Size**: Smaller files are easier to understand and navigate
+4. **Code Reuse**: Utility functions are properly separated for reuse
+5. **Testing**: Smaller modules make unit testing easier
+
+## Testing Strategy
+
+1. Create unit tests for each module to ensure core functionality
+2. Implement integration tests for interactions between modules
+3. Perform end-to-end tests of the application to verify overall functionality
+4. Use TypeScript to catch any type errors during compilation
