+++ lib/models/gear/index.ts
@@ -1,0 +1,9 @@
+// Import and re-export
+import { Gear } from './Gear';
+export { Gear };
+
+// Export types and utility functions
+export * from './types';
+export * from './processing';
+export * from './connections';
+export * from './examples';
