+++ lib/models/index.ts
@@ -1,0 +1,4 @@
+export * from './types';
+export * from './gear';
+export * from './patch';
+export * from './GearChat';
